basic working Gemini proxy
This is the first commit of Perez. It is a basically functional Gemini
over HTTP proxy, intended for personal use on a single machine, which
will eventually develop into a full-featured Gemini client.

The current version successfully proxies to Gemini and downloads
resources up to 96 KiB in size, converting gemtext to HTML. Pre elements
and links are supported, and JavaScript is used to rewrite gemini://
links to use the proxy.