improvements to HTML rendering
The Gemtext parser is now cleaner and supports all of the
elements in the Gemtext specification. I also added some CSS
to render the generated HTML in a more Gemini-like way.
Beautification of the HTML template and the CSS will be the next
step, as the pages do still look very harsh.

The markup generated is slightly idiosyncratic HTML, but it is
easy to generate from Gemtext, and it should be easy to round-trip
back to Gemtext as well. That property will be important when I
add the WYSIWYG editor.